requirements.txt
pandas
numpy
requests
geopandas
shapely
scikit-learn
xgboost
matplotlib
python-dotenv
